CHROMIUM: Changes needed for backlight control on Eve-5.4
 #1: Port https://crrev.com/c/439873 to the 5.4 tree
 #2: Port https://crrev.com/c/440445 to the 5.4 tree
 #3: Port https://crrev.com/c/456118 to the 5.4 tree

The first two patches are needed to correctly enable the backlight for
Eve; they were submitted upstream in 2017 but were rolled back.

The third patch was not rolled back upstream, but the code related to it
changed quite a bit between 4.4 and 5.4: the write of
DP_EDP_PWMGEN_BIT_COUNT used to happen on every screen enable via
intel_dp_aux_set_pwm_freq (that's how it works in 4.4), but in 5.4 the
write was refactored into intel_dp_aux_calc_max_backlight and only
called once per screen setup. This patch brings back the old
functionality of intel_dp_aux_set_pwm_freq to ensure we set that
register on each screen enable, but this is at the cost of duplication
between intel_dp_aux_calc_max_backlight. Unfortunately, I don't see a
better way to do this without making the cros code diverge even further
from 4.4 and 5.4.

I'd like to submit this CL into the chromium tree, and use b/163412221
to track its upstreaming and the ultimate revert of this change.

Test:
   * `emerge-eve-kernelnext sys-kernel/chromeos-kernel-5_4`
   * `./update_kernel.sh --remote=192.168.0.22 --remote_bootarg`
   * On DUT, set options `i915.enable_dpcd_backlight=1 i915.enable_dbc=1`
     (using e.g. `/usr/share/vboot/bin/make_dev_ssd.sh`)
   * reboot
   * Change brightness on DUT
   * Allow DUT screen to fall asleep
   * Wake DUT screen, change brightness again

BUG=b:162255390
TEST=See above

Change-Id: I9eafa28226f1c6b8332fcf9730259bea05cf7975
Signed-off-by: Kevin Chowski <chowski@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2344844
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Conflicts:
	drivers/gpu/drm/i915/display/intel_dp_aux_backlight.c
[rebase510(groeck): Picked code from continuous rebase project to resolve conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>