CHROMIUM: Disable stibp on kernels with core-scheduling enabled
HT enabling also introduces STIBP mode [1] to mitigate spectrev2. There
is no reason to keep STIBP enabled, if core-scheduling is enabled. This
is because STIBP is for cross-HT branch prediction poisoning. With
core-scheduling we avoid victims and attackers running simulataneously
on a core.

This fixes a 22% regression in Speedometer2 testing. This is a Chromium
specific patch because core-scheduling is not yet upstreamed and the
upstream solution could look slightly different.

BUG=b:159627360
TEST=speedometer2

[1] https://software.intel.com/security-software-guidance/insights/deep-dive-single-thread-indirect-branch-predictors

Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I0b6cd77f9a26cb8a509c839e7aff0a13c03e1fee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2346586
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>