FROMLIST: thunderbolt: Optimize Force Power logic
Currently the "Force Power" logic uses 10 retries, each with a delay of
250 ms. Thunderbolt controllers in Ice Lake and Tiger Lake platforms are
found to complete this in the order of 3 ms or so. Since this delay
is in resume path, surplus delay is effectively affecting runtime PM
resume flows.

Decrease the granularity of the delay to 3 ms and increase the number of
retries so we wait maximum of ~1 s which is the recommended timeout.
This should make runtime resume a bit faster.

Reported-by: Dana Alkattan <dana.alkattan@intel.com>
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(am from https://patchwork.kernel.org/patch/11723815/)

BUG=b:162700590, b:165030811
TEST=No regressions for the following use cases:
- S0iX (no device attached)
- TBT NVMe hotplug
- TBT dock hotplug
- TBT dock w/ USB pen drive
- TBT dock w/ display
TEST=S0iX works w/ TBT device attached (no PCIe tunneling)

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Ic56d8af052f0aaf766d99f23cd832ff897d54f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2369474
Tested-by: Utkarsh H Patel <utkarsh.h.patel@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>