FROMLIST: thunderbolt: Add runtime PM for Software CM
This adds runtime PM support for the Software Connection Manager parts
of the driver. This allows to save power when either there is no device
attached at all or there is a device attached and all following
conditions are true:

  - Tunneled PCIe root/downstream ports are runtime suspended
  - Tunneled USB3 ports are runtime suspended
  - No active DisplayPort stream
  - No active XDomain connection

For the first two we take advantage of device links that were added in
previous patch. Difference for the system sleep case is that we also
enable wakes when something is geting plugged in/out of the Thunderbolt
ports.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(am from https://patchwork.kernel.org/patch/11723771/)

BUG=b:162700590
TEST=No regressions for the following use cases:
- S0iX (no device attached)
- TBT NVMe hotplug
- TBT dock hotplug
- TBT dock w/ USB pen drive
- TBT dock w/ display
TEST=S0iX works w/ TBT device attached (no PCIe tunneling)

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I19c2bd590c5577595765671d7faef4dc52fc6491
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2369491
Tested-by: Utkarsh H Patel <utkarsh.h.patel@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>