FROMGIT: mmc: sdhci-acpi: Fix HS400 tuning for AMDI0040
The AMD eMMC Controller can only use the tuned clock while in HS200 and
HS400 mode. If we switch to a different mode, we need to disable the
tuned clock. If we have previously performed tuning and switch back to
HS200 or HS400, we can re-enable the tuned clock.

Previously the tuned clock was not getting disabled when switching to
DDR52 which is part of the HS400 tuning sequence.

Fixes: 34597a3f60b1 ("mmc: sdhci-acpi: Add support for ACPI HID of AMD Controller with HS400")
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Acked-by: Adrian Hunter <adrian.hunter@intel.com>
Link: https://lore.kernel.org/r/20200819125832.v2.1.Ie8f0689ec9f449203328b37409d1cf06b565f331@changeid
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 61d7437ed13906984c44697970ee792ac6271a31
 https://git.kernel.org/pub/scm/linux/kernel/git/ulfh/mmc.git next)

BUG=b:157388667
TEST=Boot trembyle and dalboz and verified HS400 was working. Then
     performed tuning and verified no tuning errors were found

Change-Id: I9c138638b74466f1680aa5365cf07c362877ec31
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2372872
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@google.com>
Tested-by: Raul E Rangel <rrangel@chromium.org>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>