CHROMIUM: drm: panel: simple: Add vrefresh for KD116N21-30NV-A010
Upstream has removed the vrefresh field in commit 0425662fdf05 ("drm:
Nuke mode->vrefresh").  That CL not only adds support to calculate the
vrefresh dynamically in some cases but it also removes all users which
makes it really hard to backport (because there are more users
upstream than downstream).  For now, let's add the vrefresh back in
for KD116N21-30NV-A010.  If later someone backports the CL removing
vrefresh it will be a super obvious compile-time error.

NOTE: if we skip this patch things seem to work for the most part, but
I'm a little worried that there's some bit of code somewhere that was
relying on it.  Best to take it.

THIS PATCH SHOULDN'T BE PICKED TO NEW KERNELS--JUST DROP IT.

BUG=b:161373813
TEST=None

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I5b18b3bdd0fa4d8332a3df1470515a335837bfa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2369492
Reviewed-by: Rob Clark <robdclark@chromium.org>