FROMGIT: ASoC: hdac_hdmi: tidy up a memset()
The ARRAY_SIZE() is the number of the elements but we want to use the
number of bytes.  Fortunately, in this case the value is the same so it
doesn't affect runtime.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Link: https://lore.kernel.org/r/20200825104623.GA278587@mwanda
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 672072976bf0db6e0aca5382bcf03bc90f439923
git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-next)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:167421905
TEST=No audio regression on TGL/JSL devices