CHROMIUM: iwl7000: print PNVM version as IWL_INFO
There is a requirement to print the PNVM version that has been loaded.
We print this whenever we get the PNVM version TLV, but only if
DEBUG_FW is enabled.  Additionally, we may skip the PNVM section where
this TLV was found, due to errors in the PNVM or HW_TYPE mismatch
etc., so we should print the version that was actually loaded.

To solve this, add a new print with IWL_INFO (so we don't need
DEBUG_FW to be enabled) and do so only after the PNVM was loaded
successfully.

Change-Id: Id9f7d347f7c91f39b01177a640e5ea63f17531e3
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
(cherry picked from commit 02bfc2e075deb310ce38d29d26d2578f3d7b16b3)
iwl7000-tree: c731df16d7100a35b3049c72ac680923c6677204