ANDROID: Re-enable menus hidden by disabling MEDIA_SUPPORT_FILTER
A recent change upstream changed how MEDIA_SUPPORT_FILTER vs driver
autoselection worked, in a way that is negative for GKI. Either, you can
enable MEDIA_SUPPORT_FILTER and have implicit driver autoselection, OR
you can disable it to control more of the core media framework, but have
to manually *deselect* the drivers you don't want. This is untenable for
GKI as it will require us to stay on top of newly added DVB/TV/Radio
drivers added upstream, and we don't even want those subsystems right
now.

So, modify the bool/prompt menus back to regular bool Kconfig options
and drop the dependency on MEDIA_SUPPORT_FILTER. This still allows the
media core to be tweaked, but the drivers will all default off, instead
of on (=m).

Bug: 163582226
Change-Id: I87f72efbf6321c194982c0bff7a80d1d255522ae
Signed-off-by: Alistair Delva <adelva@google.com>
[joneslee: Removed defconfig elements]
Signed-off-by: Lee Jones <lee.jones@linaro.org>