CHROMIUM: config: Switch proximity sensor SX9310 to module
There's no reason for this to be builtin.  Userspace should be able to
take advantage of this sensor as soon as it shows up after boot.  As
some evidence that userspace handles this, it's already a module on
x86 machines.  As further evidence, the daemon that looks at this is
powerd which generally handles udev events, and in fact you can see
"SarWatcher::OnUdevEvent" in the powerd code.

Making this a module gives us a bit of a bootperf boost (~250 ms in
the kernel) plus makes the kernel smaller (so bootloader should be
able to load it faster).  NOTE: we could get the bootperf boost by
also making the probe async but in general it's good to keep as many
things in modules as possible to keep the kernel small anyway.

This is:
  echo 'CONFIG_SX9310=m' >> \
    chromeos/config/arm64/chromiumos-arm64.flavour.config
  echo 'CONFIG_SX9310=m' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
  ./chromeos/scripts/kernelconfig olddefconfig

BUG=b:166841365
TEST=Boot is faster; powerd logs still show it finding the sensor

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I0d100286bf7ef155d88c2310cbf52dd1895d3b09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2388223
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Daniel Campello <campello@chromium.org>
(cherry picked from commit 8809b414e5aaba820b333b34532d22a3566bb98a)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2391440