CHROMIUM: arm64: dts: qcom: sc7180-pompom shouldn't inherit trogdor-r1
Let's make pompom.dtsi its own thing.  This copies some stuff from the
trogdor-r1 and modifies it as we had done in the pompom.dts.

This should have no functional change.  Indeed decompiling the dtb
file shows that it doesn't.

NOTE: at some point we might want to promote the "i2c-hid" fragment to
the main trogdor-dtsi if it turns out to be more common, or at least
make it a fragment.  For now we'll just duplicate.

NOTE ALSO: The goal of this patch is no change.  A future patch will
adjust pin names and add pompom-r1.

BUG=b:166812622, b:163613554
TEST=No changes at all.

Change-Id: I2f0c23af9d1ca297b4ed2816cb8fed2950198785
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2398982
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>