CHROMIUM: arm64: dts: qcom: Add sc7180-pompom-r1
Pompom has two kernel-visible changes from pompom-r0:
* TP_INT_ODL change from GPIO58 to GPIO0
* DMIC_SEL change from GPIO18 to GPIO86

We're not modeling DMIC_SEL in the device tree yet (and it's also not
on any trogdor variants except pompom), so that one is easy.

For TP_INT_ODL it's a bit zanier.  TP_INT_ODL is the new name for what
used to be called TRACKPAD_INT_1V8_ODL.  GPIO0 is the new official
place for all (or nearly all?) boards going forward.  That means that
the correct thing to do is to move / rename this in the main
trogdor-dtsi file and then have all "older" revisions (pretty much
everything at this point) override it.

This change is at least moderately ugly because trogdor.dtsi is now
upstream but trogdor-r0 is not (and probably never will by) nor is
pompom (not quite ready).  For now I think we'll have to do this as a
CHROMIUM patch and sort it out once we have an actual upstream board
that moves TP_INT_ODL.

NOTE: pin-names are copy-pasted from EE-provided spreadsheet and
sanity-checked.

For testing, I did this to reverse-engineer:
  cd /build/trogdor/var/cache/portage/sys-kernel/chromeos-kernel-5_4/arch/arm64/boot/dts/qcom
  for dtb in sc7180-trogdor-*.dtb; do
    dtc -I dtb -O dts ${dtb} > ${PATH_TO}/${dtb}.dts;
  done

BUG=b:166812622, b:163613554
TEST=Build and boot
TEST=Reverse-compile before and after and see dts files seem sane.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: If0b6de0cdeb1f66d7d0066e4d0ddc5319b56f0a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2398983
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
(cherry picked from commit 013f1e9970de20b6c9c8b7c4e16164fd6999884a)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2402804
Reviewed-by: AJ Cheng <aj_cheng@compal.corp-partner.google.com>
Tested-by: AJ Cheng <aj_cheng@compal.corp-partner.google.com>