FROMGIT: spi: spi-geni-qcom: Don't program CS_TOGGLE again and again
We always toggle the chip select manually in spi-geni-qcom so that we
can properly implement the Linux API.  There's no reason to program
this to the hardware on every transfer.  Program it once at init and
be done with it.

This saves some part of a microsecond of overhead on each transfer.
While not really noticeable on any real world benchmarks, we might as
well save the time.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20200912140730.2.I33e571179986850b4ec17042e813d0b08fb1b9c1@changeid
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 14ac4e049dc1183440960f177b60b54357e54d90
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/spi.git for-next)

BUG=b:159849273
TEST=flashrom -p ec -r /tmp/foo.bin

Change-Id: I33e571179986850b4ec17042e813d0b08fb1b9c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2410463
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>