UPSTREAM: drm/rockchip: cdn-dp: fix sign extension on an int multiply for a u64 result
The variable bit_per_pix is a u8 and is promoted in the multiplication
to an int type and then sign extended to a u64. If the result of the
int multiplication is greater than 0x7fffffff then the upper 32 bits will
be set to 1 as a result of the sign extension. Avoid this by casting
tu_size_reg to u64 to avoid sign extension and also a potential overflow.

Fixes: 1a0f7ed3abe2 ("drm/rockchip: cdn-dp: add cdn DP support for rk3399")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Signed-off-by: Heiko Stuebner <heiko@sntech.de>
Link: https://patchwork.freedesktop.org/patch/msgid/20200915162049.36434-1-colin.king@canonical.com
(cherry picked from commit ce0cb93a5adb283f577cd4661f511047b5e39028)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:191366271
TEST=Tested with single/multiple displays, suspend/resume, hotplugs on volteer/zork/trogdor. Build tested on others

Backported using forklift.py

Change-Id: I763f2d02e0f26d4f6945ed73b5adccf61d48b464