CHROMIUM: mm: totmaps: split PSS into components
This gives us separate numbers for PSS anon, file, and shmem.
We need this to measure types of usage by process group.

This is a port of the 4.4 version, we're still using totmaps instead
of smaps_rollup in some circumstances because of permissions issues
with smaps_rollup.

BUG=b:148899837, b:162698408
TEST=chromeos-pgmem returns non-zero values

[5.4: SR: sum up the new pss components properly]
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Change-Id: Ibbc4caff0a03e2356c74598fe4a4e01909131d28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2429965
Reviewed-by: Kuo-Hsin Yang <vovoy@chromium.org>
(cherry picked from commit f277495e97f99d0f778ec550f75628a85522083c)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2437133