CHROMIUM: sched/core: Make threads in a process share a core
Previously, we did not allow threads in a process to share a core. However,
this does make a lot of sense. Threads in a process share the same virtual
memory space so they can attack each other anyway. There's not much point in
losing the performance. Let us make threads in a process share the same core.

BUG=b:152605392
TEST=ARC Audio tests
Signed-off-by: Joel Fernandes <joelaf@google.com>
Change-Id: I2c342f2512ea18ceca880d9492070c39d9c0be4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2434732