CHROMIUM: sched/coresched: consider forced idle task for sched_slice
It is seen that sometimes a task run for too long if another sibling is
force-idled and there is only a single task running on the current RQ. In this
case, don't let the sibling run for too long without entering the schedule loop.

Patch is from Vineeth and was not sent yet to mailing list as the series is
being prepared, however it shows improvement in tests.

TEST=power_VideoCall.fast
BUG=b:152605392

Signed-off-by: Vineeth Remanan Pillai <viremana@linux.microsoft.com>
Signed-off-by: Joel Fernandes <joelaf@google.com>

Change-Id: I4925730ac9feddcdd2b51ea14de84bd9cce40014
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2441259
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Joel Fernandes <joelaf@google.com>
Commit-Queue: Joel Fernandes <joelaf@google.com>