UPSTREAM: drm/amd/display: Do not warn NULL dc_sink if forcing connector
[why]
There is a DRM_ERROR when the dc_sink is NULL and
there should not be this warning when the connector
is forced.

[how]
Do not warn if dc_sink is NULL if the connector
is forced.

Signed-off-by: Victor Lu <victorchengchi.lu@amd.com>
Reviewed-by: Nicholas Kazlauskas <Nicholas.Kazlauskas@amd.com>
Acked-by: Qingqing Zhuo <qingqing.zhuo@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit ad975f448d963d700aa9a67b9d6aa5489c00cad2)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:177892854
TEST=Tested on volteer, build-tested trogdor

Change-Id: I0a0c520f7ab2ec3105163dd4798c161733d9679e