FIXUP: FROMLIST: sched/core: Sync the min_vruntime of cores when the system enters force-idle
Just a slight optimization to the initial patch that did this. The original
patch had no bug, but we settled on doing it this way. So fix the patch to do
it this way.

Change-Id: I15d9693c327f47829bfb9545d4d095c74988ed9a
Signed-off-by: Joel Fernandes <joelaf@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2441261
(cherry picked from commit ae180b997ef1d57b7992c8ba5f4cb3fa036f8d7c)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2456082